+++ semver-project-gradle-plugin/testFunctional/kotlin/com/javiersc/semver/project/gradle/plugin/PushSemverTagTaskTest.kt
@@ -29,7 +29,8 @@ class PushSemverTagTaskTest : GradleTestKitTest() {
                         .setGitDir(remoteBuildGitDir)
                         .readEnvironment()
                         .findGitDir()
-                        .build())
+                        .build()
+                )
 
             git.repository.directory.copyRecursively(remoteBuildGitDir)